Delete util.rs, move generate_random_string to root
